backend/requirements.txt
# Python 3.10-3.12 recommended (3.13+ may lack wheels for some packages)
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
python-multipart>=0.0.9
sqlalchemy>=2.0.0,<3
passlib[bcrypt]>=1.7.4,<2
# passlib 1.7.x: pin bcrypt 3.x (incompatible with bcrypt 4.x)
bcrypt>=3.2.0,<4.0.0
python-jose[cryptography]>=3.3.0,<4
email-validator>=2.0.0,<3
python-dotenv>=1.0.0,<2
pandas>=2.0.0,<3
openpyxl>=3.1.0,<4
# 프로젝트 브리프 파일에서 텍스트 추출 (Word/Excel/PPT/PDF 등)
python-docx>=1.0.0,<2
python-pptx>=0.6.21,<2
pypdf>=4.0.0,<6
xlrd>=2.0.0,<3
odfpy>=1.4.0,<2
striprtf>=0.0.26,<1
numpy>=1.26.0,<3
scikit-learn>=1.4.0,<2
xgboost>=2.0.0,<3
matplotlib>=3.8.0,<4
joblib>=1.3.0,<2
psutil>=5.9.0,<6
torch
lightning
pytorch-forecasting
# Optional: AI chat (OpenAI / Gemini / Ollama OpenAI-compatible)
openai>=1.51.0,<2
google-generativeai>=0.8.0,<0.9

# Colab 유사 노트북 (선택): NOTEBOOK_ENABLED=1 일 때 사용
jupyterlab>=4.0.0,<5
